pyproject.toml
[tool.poetry]
name = "uart"
version = "0.1.0"
description = ""
authors = ["juan pablo saucedo <juan.s@ensolvers.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
